Add method and update comments

+++ mechtatel-common-utils/src/main/java/com/github/maeda6uiui/mechtatel/common/utils/MttResourceFileUtils.java
@@ -14,15 +14,20 @@ public class MttResourceFileUtils {
     /**
      * Extracts a file from a JAR and writes it out to a temporary file.
      *
-     * @param clazz    Class to call {@link Class#getResourceAsStream(String)}
-     * @param filepath Filepath of the file
+     * @param clazz          Class to call {@link Class#getResourceAsStream(String)}
+     * @param filepath       Filepath of the file
+     * @param tempFilePrefix Prefix for the temporary file
+     * @param deleteOnExit   Deletes the temporary file on exit if true
      * @return Temporary file representing the extracted file
      * @throws IOException If it fails to extract the file
      */
-    public static File extractFile(Class<?> clazz, String filepath) throws IOException {
+    public static File extractFile(
+            Class<?> clazz, String filepath, String tempFilePrefix, boolean deleteOnExit) throws IOException {
         try (var bis = new BufferedInputStream(Objects.requireNonNull(clazz.getResourceAsStream(filepath)))) {
-            File tempFile = File.createTempFile("mtt", ".tmp");
-            tempFile.deleteOnExit();
+            File tempFile = File.createTempFile(tempFilePrefix, ".tmp");
+            if (deleteOnExit) {
+                tempFile.deleteOnExit();
+            }
 
             try (var bos = new BufferedOutputStream(new FileOutputStream(tempFile))) {
                 bis.transferTo(bos);
@@ -30,6 +35,18 @@ public class MttResourceFileUtils {
 
             return tempFile;
         }
+    }
+
+    /**
+     * Extracts a file from a JAR and writes it out to a temporary file.
+     *
+     * @param clazz    Class to call {@link Class#getResourceAsStream(String)}
+     * @param filepath Filepath of the file
+     * @return Temporary file representing the extracted file
+     * @throws IOException If it fails to extract the file
+     */
+    public static File extractFile(Class<?> clazz, String filepath) throws IOException {
+        return extractFile(clazz, filepath, "mtt", true);
     }
 
     /**
@@ -59,13 +76,18 @@ public class MttResourceFileUtils {
 
     /**
      * Loads a native library contained in a JAR.
+     * This method first extracts the native library from inside a JAR to a temporary file,
+     * and then loads it with {@link System#load(String)}.
+     * This method does not delete the temporary file because {@link File#deleteOnExit()} does not work
+     * on Windows probably because Windows still locks the DLL file even at the moment of JVM shutdown.
+     * The temporary file is created with a prefix of "mttnatives" under the OS-dependent temp directory.
      *
      * @param clazz    Class to call {@link Class#getResourceAsStream(String)}
      * @param filepath Filepath of the native library
      * @throws IOException If it fails to extract the file
      */
     public static void loadNativeLib(Class<?> clazz, String filepath) throws IOException {
-        File tempFile = extractFile(clazz, filepath);
+        File tempFile = extractFile(clazz, filepath, "mttnatives", false);
         System.load(tempFile.getAbsolutePath());
     }
 }
